+++ .gitmodules
@@ -1,6 +1,3 @@
-[submodule "packages/contracts/lib/forge-std"]
-	path = packages/contracts/lib/forge-std
+[submodule "packages/foundry/lib/forge-std"]
+	path = packages/foundry/lib/forge-std
 	url = https://github.com/foundry-rs/forge-std
-[submodule "packages/contracts/lib/openzeppelin-contracts"]
-	path = packages/contracts/lib/openzeppelin-contracts
-	url = https://github.com/openzeppelin/openzeppelin-contracts
