+++ .prettierrc.json
@@ -1,10 +1,10 @@
 {
   "trailingComma": "es5",
   "semi": false,
-  "singleQuote": true,
+  "singleQuote": false,
   "jsxSingleQuote": true,
   "printWidth": 120,
   "bracketSameLine": true,
   "useTabs": false,
-  "tabWidth": 2
+  "tabWidth": 4
 }
