--- packages/contracts/foundry.toml
@@ -1,6 +1,0 @@
-[profile.default]
-src = 'src'
-out = 'out'
-libs = ['lib']
-
-# See more config options https://github.com/foundry-rs/foundry/tree/master/config
