--- packages/contracts/remappings.txt
@@ -1,3 +1,0 @@
-forge-std/=lib/forge-std/src/
-ds-test/=lib/forge-std/lib/ds-test/src/
-@openzeppelin/=lib/openzeppelin-contracts/
