+++ packages/foundry/remappings.txt
@@ -1,0 +1,2 @@
+forge-std/=lib/forge-std/src/
+ds-test/=lib/forge-std/lib/ds-test/src/
