+++ packages/hardhat/package.json
@@ -1,0 +1,37 @@
+{
+    "name": "nexth-contracts",
+    "version": "0.1.0",
+    "license": "MIT",
+    "scripts": {
+        "build": "hardhat compile",
+        "deploy": "npx hardhat run scripts/deploy.ts",
+        "deploy:nft": "npx hardhat run scripts/nft.ts",
+        "coverage": "hardhat coverage",
+        "test": "REPORT_GAS=true hardhat test"
+    },
+    "devDependencies": {
+        "@ethersproject/abi": "^5.7.0",
+        "@ethersproject/providers": "^5.7.2",
+        "@nomicfoundation/hardhat-chai-matchers": "^1.0.0",
+        "@nomicfoundation/hardhat-network-helpers": "^1.0.0",
+        "@nomicfoundation/hardhat-toolbox": "^2.0.0",
+        "@nomiclabs/hardhat-ethers": "^2.2.3",
+        "@nomiclabs/hardhat-etherscan": "^3.1.7",
+        "@openzeppelin/contracts": "^5.0.0",
+        "@typechain/ethers-v5": "^10.1.0",
+        "@typechain/hardhat": "^6.1.2",
+        "@types/chai": "^4.2.0",
+        "@types/dotenv": "^8.2.0",
+        "@types/mocha": ">=9.1.0",
+        "@types/node": ">=12.0.0",
+        "chai": "^4.2.0",
+        "dotenv": "^16.3.1",
+        "ethers": "5.7.2",
+        "hardhat": "^2.17.4",
+        "hardhat-gas-reporter": "^1.0.9",
+        "solidity-coverage": "^0.8.5",
+        "ts-node": "^10.9.1",
+        "typechain": "^8.3.1",
+        "typescript": "^5.2.2"
+    }
+}
