+++ packages/hardhat/scripts/nft.ts
